requirements.txt
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
cycler==0.12.1
idna==3.8
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.4.3
multitasking==0.0.11
nltk==3.6.3
numpy==1.21.2
oauthlib==3.2.2
pandas==1.3.3
pillow==10.4.0
praw==7.7.1
prawcore==2.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==0.19.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==1.3.1
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
textblob==0.15.3
threadpoolctl==3.5.0
tqdm==4.66.5
tweepy==4.14.0
update-checker==0.18.0
urllib3==1.26.20
websocket-client==1.8.0
yfinance==0.1.63
